+++ CHANGELOG.md
@@ -11,8 +11,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 ### Changed
+
+### Removed
+
+## [0.20.0](https://github.com/lightstep/opentelemetry-prometheus-sidecar/releases/tag/v0.20.0) - 2021-03-22
+
+### Changed
 - Uses Prometheus `/api/v1/status/config` endpoint to read the Prometheus 
   config, to automatically determine the full set of scrape intervals. (#162)
+- The startup timeout is raised to 10 minutes. (#166)
 
 ### Removed
 - The `--prometheus.scrape-interval` option is ignored. (#162)